Fix build failure by restoring less variable
The warning variable in variables.less sees to have been removed in
commit d8fc98b during PR #592 (by mistake?).

Mysteriously, the builds passed for this PR, but it was after the merge
of the next PR #591 that the build started failing. Furthermore, this
second PR #591 was also passing in isolation, but a local experiment
rebasing it on top of the previous one shows the build failure.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>