Update feature toggles to respect env variables
If `MENDER_HOSTED` is set, fature toggles do not respect the environment
variables and rely on the tenants' capabilities (enabled add-ons) only.
This makes impossible to disable a feature in Hosted Mender when the
backend already supports a new add-on.

Make the feature toggles always respect the environment variables.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>