MEN-4411: Show dates formatted according to user locale
This changes the date-format string to rely on the locale, as defined by the
browser's locale settings, and can further be modified through a menu in 'User
Management', to override the standard.

The option is saved for the next session, as a standard user setting.

Changelog: Enable the user to pick the time locale, through the 'User
Management' interface. The default is to rely on the browser settings, but is
user modifiable through the drop-down menu.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>