fix: move cookie-based JWT tokens to the local storage
If the log in happens through SSO (e.g., OAuth 2.0 or SAML 2) the
backend sets the JWT token in a cookie. The UI is then responsible for
moving it to the local storage, as happens with the JWT token on a
standard email/password log in. This fixes the expiration issues with
cookie-based JWT tokens.

Changelog: None
Ticket: MEN-6849

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>